requirements.txt
mlflow-skinny
azureml-mlflow
scikit-learn
pandas